requirements.txt
asgiref==3.4.1
attrs==21.4.0
black==21.12b0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.11.0
django-enum-choices==2.1.3
django-enumchoicefield==2.0.0
django-filter==21.1
django-rest-framework==0.1.0
django-rest-passwordreset==1.2.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-extra-fields==3.2.1
drf-writable-nested==0.6.3
dropbox==11.25.0
flake8==4.0.1
generics==3.4.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
msgpack==1.0.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
ply==3.11
proto-plus==1.19.8
protobuf==3.19.3
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.1.1
pytest-django==4.5.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stone==3.2.1
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
